chore: fix hook generation

+++ snapcraft/parts/setup_assets.py
@@ -336,7 +336,8 @@ def _write_hook_wrapper(hook_name: str, wrapper_path: Path) -> None:
     :param hook_name: name of the hook
     :param wrapper_path: file path of hook wrapper
     """
-    wrapper_path.unlink(missing_ok=True)
+    if wrapper_path.exists():
+        return
 
     with open(wrapper_path, "w+", encoding="utf-8") as wrapper_file:
         print("#!/bin/sh", file=wrapper_file)
